+++ core/modules/modSamlConnector.class.php
@@ -60,7 +60,7 @@ class modSamlConnector extends DolibarrModules {
         $this->editor_url = 'https://www.atm-consulting.fr/';
 
         // Possible values for version are: 'development', 'experimental', 'dolibarr', 'dolibarr_deprecated' or a version string like 'x.y.z'
-        $this->version = '1.0.0';
+        $this->version = '1.0.1';
         // Url to the file with your last numberversion of this module
         //$this->url_last_version = 'http://www.example.com/versionmodule.txt';