+++ lib/onelogin/php-saml/src/Saml2/Utils.php
@@ -245,6 +245,13 @@ class Utils
                 if ($heads) {
                     $key = "-----BEGIN RSA PRIVATE KEY-----\n".chunk_split($key, 64, "\n")."-----END RSA PRIVATE KEY-----\n";
                 }
+            } else if (strpos($key, '-----BEGIN ENCRYPTED PRIVATE KEY-----') !== false) {
+                $key = Utils::getStringBetween($key, '-----BEGIN ENCRYPTED PRIVATE KEY-----', '-----END ENCRYPTED PRIVATE KEY-----');
+                $key = str_replace(' ', '', $key);
+
+                if ($heads) {
+                    $key = "-----BEGIN ENCRYPTED PRIVATE KEY-----\n".chunk_split($key, 64, "\n")."-----END ENCRYPTED PRIVATE KEY-----\n";
+                }
             } else {
                 $key = str_replace(' ', '', $key);