+++ ChangeLog.md
@@ -1,5 +1,6 @@
 # CHANGELOG SAMLCONNECTOR FOR [DOLIBARR ERP CRM](https://www.dolibarr.org)
 
 ## 1.0
+- FIX : Missing Encrypted Key Type *05/01/2023* 1.0.1
 
 Initial version